Move ColorPicker shaders to ColorPickerShape class
